ibkr_show_backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
elasticsearch>=8,<9
pytest
httpx
